django/requirements.txt
asgiref==3.4.1
autoflake==1.4
black==21.9b0
click==8.0.1
Django==3.2.7
django-cors-headers==3.8.0
django-filter==21.1
djangorestframework==3.12.4
flake8==3.9.2
isort==5.9.3
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
regex==2021.8.28
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.2
uWSGI==2.0.19.1
